Move bases classes to utils submodule
