+++ qc_firmware/scripts/ptctestsuite/__init__.py
@@ -1,3 +1,3 @@
-from ptctestsuite.qc_record import qc_record
-from ptctestsuite.qc_result import qc_result
-from ptctestsuite.test_base import test_base
+from ptctestsuite.utils import qc_record
+from ptctestsuite.utils import qc_result
+from ptctestsuite.utils import test_base
