+++ qc_firmware/scripts/ptctestsuite/ptctests/README.md
@@ -1,0 +1,3 @@
+## PTC Tests
+
+This folder contains the collection of scripts to be run on the PTC itself.
