+++ qc_firmware/scripts/ptctestsuite/utils/__init__.py
@@ -1,0 +1,3 @@
+from ptctestsuite.utils.qc_record import qc_record
+from ptctestsuite.utils.qc_result import qc_result
+from ptctestsuite.utils.test_base import test_base
