+++ qc_firmware/scripts/ptctestsuite/utils/qc_record.py
