+++ qc_firmware/scripts/ptctestsuite/utils/qc_result.py
